// msp/src/app/components/msp/model/spa-env-response.interface.ts
export interface ISpaEnvResponse {
    SPA_ENV_MSP_MAINTENANCE_FLAG: string;
    SPA_ENV_MSP_MAINTENANCE_MESSAGE: string;
    SPA_ENV_ACL_MAINTENANCE_FLAG: string;
    SPA_ENV_ACL_MAINTENANCE_MESSAGE: string;
    SPA_ENV_PACUTOFF_MAINTENANCE_FLAG?: string;
    SPA_ENV_PACUTOFF_MAINTENANCE_MESSAGE?: string;
    SPA_ENV_PACUTOFF_MAINTENANCE_START: string;
    SPA_ENV_NOW: string;
    SPA_ENV_PACUTOFF_MAINTENANCE_END: string;
    SPA_ENV_SUPPBEN_MAINTENANCE_FLAG?: string;
    SPA_ENV_SUPPBEN_MAINTENANCE_MESSAGE?: string;
    SPA_ENV_SUPPBEN_MAINTENANCE_START?: string;
    SPA_ENV_SUPPBEN_MAINTENANCE_END?: string;
}